+++ jackson/src/main/scala/com/twitter/finatra/json/internal/caseclass/validation/validators/MaxValidator.scala
@@ -38,7 +38,7 @@ private[finatra] class MaxValidator(
       case numberValue: Number =>
         validationResult(numberValue)
       case _ =>
-        throw new IllegalArgumentException(s"Class [${value.getClass}] is not supported")
+        throw new IllegalArgumentException(s"Class [${value.getClass}] is not supported by ${this.getClass}")
     }
   }